turn constants file into theme and utils files
