+++ movies/lib/src/app.dart
@@ -3,22 +3,14 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:movies/constants.dart';
+import 'package:movies/utils.dart';
 import 'package:movies/src/blocs/genres_cubit.dart';
 import 'package:movies/src/blocs/movie_details_cubit.dart';
 import 'package:movies/src/blocs/movie_search_bloc.dart';
 import 'package:movies/src/blocs/selected_movie_bloc.dart';
 import 'package:movies/src/pages/details/details_page.dart';
 import 'package:movies/src/pages/search/search_page.dart';
-
-final themeData = ThemeData(
-  scaffoldBackgroundColor: bgColor,
-  colorScheme: const ColorScheme.dark(
-    primary: accentColor,
-    secondary: accentColor,
-    background: bgColor,
-  ),
-);
+import 'package:movies/theme.dart';
 
 /// The Widget that configures your application.
 class MyApp extends StatelessWidget {
